backend/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
python-multipart
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langchain-mistralai==0.1.0
langchain-chroma==1.0.0
pypdf==6.3.0
sentence-transformers==5.1.2
chromadb==1.3.5
python-dotenv
